feat: debitos na conta do cliente quando pedido é aprovado

+++ db/migrate/20220630152524_create_debits.rb
@@ -1,8 +1,8 @@
 class CreateDebits < ActiveRecord::Migration[7.0]
   def change
     create_table :debits do |t|
-      t.integer :real_amount
-      t.integer :rubi_amount
+      t.decimal :real_amount
+      t.decimal :rubi_amount
       t.references :exchange_rate, null: false, foreign_key: true
       t.references :client, null: false, foreign_key: true
       t.references :order, null: false, foreign_key: true
